requirements.txt
Pillow>=10.0.0
tqdm>=4.0.0
rich>=13.0.0
questionary>=2.0.0
rembg>=2.0.0
onnxruntime>=1.15.0
